feat: Chapter, Page Draggable 삭제 Droppable 영역 추가
