+++ src/main/react/book_write/public/index.html
@@ -40,7 +40,7 @@
   <title>write</title>
 </head>
 <body class="bg-gray-50">
-<div id="root" class="flex"></div>
+<div id="root" class="flex relative"></div>
 
 </body>
 </html>
